# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "News Aggregator API",
    "version": "1.0.0",
    "description": "Public API for aggregating news from multiple providers. This API returns curated news articles for users.",
    "contact": {
      "name": "News Team",
      "email": "shanmuka089@gmail.com",
      "url": "https://example.com/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/news-aggregator/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080/news-aggregator/api/v1",
      "description": "Staging"
    }
  ],
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "Users Management"
        ],
        "summary": "Register User",
        "description": "Registers a new user in the system",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              },
              "example": {
                "username": "John Doe",
                "email": "johndoe@gmail.com",
                "password": "Secured@123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                      "message": "Validation failed",
                      "errorMessage": "Invalid input data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                      "message": "Unauthorized access",
                      "errorMessage": "You must be logged in to access this resource"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                      "message": "Access denied",
                      "errorMessage": "You do not have permission to perform this action"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                      "message": "An unexpected error occurred",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                      "message": "Service is currently unavailable",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [],
        "security": []
      },
      "get": {
        "tags": [
          "Users Management"
        ],
        "summary": "Get User",
        "description": "Retrieves user details",
        "requestBody": {},
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                      "message": "Validation failed",
                      "errorMessage": "Invalid input data"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                      "message": "Unauthorized access",
                      "errorMessage": "You must be logged in to access this resource"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                      "message": "Access denied",
                      "errorMessage": "You do not have permission to perform this action"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                      "message": "An unexpected error occurred",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                },
                "examples": {
                  "BadRequest": {
                    "value": {
                      "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                      "message": "Service is currently unavailable",
                      "errorMessage": "Please try again later"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [],
        "security": []
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users Management"
        ],
        "summary": "Login User",
        "description": "Logs in a user and returns an authentication token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              },
              "example": {
                "email": "johndoe@gmail.com",
                "password": "Secured@123"
              }
            }
          },
          "responses": {
            "200": {
              "description": "User logged in successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                        "message": "Validation failed",
                        "errorMessage": "Invalid input data"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                        "message": "Unauthorized access",
                        "errorMessage": "You must be logged in to access this resource"
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Forbidden",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                        "message": "Access denied",
                        "errorMessage": "You do not have permission to perform this action"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                        "message": "An unexpected error occurred",
                        "errorMessage": "Please try again later"
                      }
                    }
                  }
                }
              }
            },
            "503": {
              "description": "Service Unavailable",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                        "message": "Service is currently unavailable",
                        "errorMessage": "Please try again later"
                      }
                    }
                  }
                }
              }
            }
          },
          "parameters": [],
          "security": []
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Users Management"
        ],
        "summary": "Logout User",
        "description": "Logs out a user and invalidates the authentication token",
        "requestBody": {},
          "responses": {
            "200": {
              "description": "User logged out successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_BAD_REQUEST",
                        "message": "Validation failed",
                        "errorMessage": "Invalid input data"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_UNAUTHORIZED",
                        "message": "Unauthorized access",
                        "errorMessage": "You must be logged in to access this resource"
                      }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Forbidden",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_FORBIDDEN",
                        "message": "Access denied",
                        "errorMessage": "You do not have permission to perform this action"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_INTERNAL_SERVER_ERROR",
                        "message": "An unexpected error occurred",
                        "errorMessage": "Please try again later"
                      }
                    }
                  }
                }
              }
            },
            "503": {
              "description": "Service Unavailable",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorDto"
                  },
                  "examples": {
                    "BadRequest": {
                      "value": {
                        "errorCode": "NEWS_API_CLIENT_SERVICE_UNAVAILABLE",
                        "message": "Service is currently unavailable",
                        "errorMessage": "Please try again later"
                      }
                    }
                  }
                }
              }
            }
          },
          "parameters": [],
          "security": []
        }
      },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Error code representing the type of error",
            "example": "NEWS_API_CLIENT_BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "A human-readable message providing more details about the error",
            "example": "Validation failed"
          },
          "errorMessage": {
            "type": "string",
            "description": "Detailed error message, if available",
            "example": "Invalid input data"
          }
        }
      },
...
      "ResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ResponseStatus"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResponseStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Response code indicating success or failure",
            "example": "SUCCESS"
          }
        }
      },
...
    }
  }
}